+++ tests/EnumKeyTest.php
@@ -4,6 +4,7 @@ namespace BenSampo\Enum\Tests;
 
 use PHPUnit\Framework\TestCase;
 use BenSampo\Enum\Rules\EnumKey;
+use BenSampo\Enum\Tests\Enums\UserType;
 
 class EnumKeyTest extends TestCase
 {
